+++ src/main/java/de/is24/deadcode4j/Utils.java
@@ -68,19 +68,19 @@ public final class Utils {
         super();
     }
 
+
     /**
-     * Adds the given value to a <code>Set</code> being mapped by the specified key - and creates a new <code>Set</code>
-     * if the map does not already contain one.
+     * Retrieves an existing <code>Set</code> being mapped by the specified key or puts a new one into the map.
      *
      * @since 1.4
      */
-    public static <K, V> void addToMappedSet(Map<K, Set<V>> map, K key, V value) {
+    public static <K, V> Set<V> getOrAddMappedSet(Map<K, Set<V>> map, K key) {
         Set<V> values = map.get(key);
         if (values == null) {
             values = newHashSet();
             map.put(key, values);
         }
-        values.add(value);
+        return values;
     }
 
 }