+++ src/test/java/ClassImplementingCloneable.java
@@ -1,0 +1,3 @@
+@SuppressWarnings("UnusedDeclaration")
+public class ClassImplementingCloneable implements Cloneable {
+}