+++ src/test/java/de/is24/deadcode4j/analyzer/An_AnnotationsAnalyzer.java
@@ -1,51 +1,57 @@
 package de.is24.deadcode4j.analyzer;
 
+import de.is24.deadcode4j.Analyzer;
 import de.is24.deadcode4j.CodeContext;
-import org.junit.Before;
 import org.junit.Test;
 
 import java.util.Map;
 
-import static com.google.common.collect.Iterables.getOnlyElement;
+import static com.google.common.collect.Iterables.concat;
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.*;
+import static org.hamcrest.Matchers.containsInAnyOrder;
+import static org.hamcrest.Matchers.is;
 
 public final class An_AnnotationsAnalyzer extends AnAnalyzer {
 
-    private AnnotationsAnalyzer objectUnderTest;
-
-    @Before
-    public void setUpObjectUnderTest() {
-        this.objectUnderTest = new AnnotationsAnalyzer("junit", "de.is24.deadcode4j.junit.Annotation") {
-        };
-    }
-
     @Test
     public void reportsExistenceOfClass() {
+        Analyzer objectUnderTest = new AnnotationsAnalyzer("junit", "de.is24.deadcode4j.junit.Annotation") {
+        };
         CodeContext codeContext = new CodeContext();
-        objectUnderTest.doAnalysis(codeContext, getFile("AnnotatedClass.class"));
 
-        assertThat("Should analyze one class", codeContext.getAnalyzedCode().getAnalyzedClasses(), hasSize(1));
-        assertThat(codeContext.getAnalyzedCode().getAnalyzedClasses(), contains("AnnotatedClass"));
+        objectUnderTest.doAnalysis(codeContext, getFile("A.class"));
+        assertThat(codeContext.getAnalyzedCode().getAnalyzedClasses(), containsInAnyOrder("A"));
+
+        objectUnderTest.doAnalysis(codeContext, getFile("B.class"));
+        assertThat(codeContext.getAnalyzedCode().getAnalyzedClasses(), containsInAnyOrder("A", "B"));
     }
 
     @Test
     public void reportsAnnotatedClassAsBeingUsed() {
+        Analyzer objectUnderTest = new AnnotationsAnalyzer("junit", "de.is24.deadcode4j.junit.Annotation", "java.lang.Deprecated") {
+        };
         CodeContext codeContext = new CodeContext();
+
         objectUnderTest.doAnalysis(codeContext, getFile("AnnotatedClass.class"));
+        objectUnderTest.doAnalysis(codeContext, getFile("DeadServlet.class"));
 
         Map<String, ? extends Iterable<String>> codeDependencies = codeContext.getAnalyzedCode().getCodeDependencies();
-        assertThat("Should have reported class as being in use", codeDependencies.size(), is(1));
-        assertThat(getOnlyElement(codeDependencies.values()), hasItem("AnnotatedClass"));
+        assertThat("Should have reported some dependencies!", codeDependencies.size(), is(1));
+        assertThat(concat(codeDependencies.values()), containsInAnyOrder("AnnotatedClass", "DeadServlet"));
     }
 
     @Test
     public void doesNotReportUnannotatedClassAsBeingUsed() {
+        Analyzer objectUnderTest = new AnnotationsAnalyzer("junit", "de.is24.deadcode4j.junit.Annotation") {
+        };
         CodeContext codeContext = new CodeContext();
-        objectUnderTest.doAnalysis(codeContext, getFile("SingleClass.class"));
+
+        objectUnderTest.doAnalysis(codeContext, getFile("AnnotatedClass.class"));
+        objectUnderTest.doAnalysis(codeContext, getFile("DeadServlet.class"));
 
         Map<String, ? extends Iterable<String>> codeDependencies = codeContext.getAnalyzedCode().getCodeDependencies();
-        assertThat("Should NOT have reported class as being in use", codeDependencies.size(), is(0));
+        assertThat("Should have reported some dependencies!", codeDependencies.size(), is(1));
+        assertThat(concat(codeDependencies.values()), containsInAnyOrder("AnnotatedClass"));
     }
 
 }