+++ src/test/java/de/is24/deadcode4j/analyzer/superclass/SuperClassMarkingCodeAsLive.java
@@ -1,4 +1,0 @@
-package de.is24.deadcode4j.analyzer.superclass;
-
-public class SuperClassMarkingCodeAsLive {
-}